db exceptions mgmt and alerts

+++ app/models/comments.php
@@ -17,42 +17,24 @@ class comments extends Model
         ];
     }
 
-    public function getComments(){
-        $result = $this->select(['a.*'])
-        ->join('posts b', 'a.postId=b.id', 'LEFT')
-        ->join('user c', 'c.id=a.userId', 'LEFT')
-        ->where([['c.active', 1]])
-        ->orderBy([['a.created_at', 'DESC']])
-        ->get();
-        return $result;
-    }
-    
-    public function getUserComments($params){
-
-        $userId = $params[2];
-        $result = $this->select(['a.*, b.*, c.username, d.profilePic'])
-        ->join('posts b', 'a.postId=b.id')
-        ->join('user c', 'b.userId=c.id')
-        ->join('userinfo d', 'c.id=d.userId')
-        ->where([['a.userId', $userId]])
-        ->orderBy([['b.created_at', 'DESC']])
-        ->get();
-        return $result;
-    }
-
     public function addComment($data, $params)
     {
         $userId = $params[2];
         $postId = $params[3];
-        $this->values = [
-            $userId,
-            $postId,
-            $data["content"],
-            NULL
-        ];
-        return $this->insert();
+        if (!empty($data['content']) && !empty($userId) && !empty($postId)) {
+            $this->values = [
+                $userId,
+                $postId,
+                $data["content"],
+                NULL
+            ];
+            return $this->insert();
+        } else {
+            echo json_encode(["r" => 'e']);
+            return false;
+        }
     }
-    
+
     public function replyComment($data, $params)
     {
         $userId = $params[2];
@@ -69,7 +51,37 @@ class comments extends Model
 
     public function deleteComment($data)
     {
-        $this->where([['id', $data['commentId']]]);
-        return $this->delete();
+        if (!empty($data['commentId'])) {
+            $this->where([['id', $data['commentId']]]);
+            return $this->delete();
+        } else {
+            echo json_encode(["r" => 'e']);
+            return false;
+        }
+    }
+
+    public function getComments()
+    {
+        $result = $this->select(['a.*'])
+            ->join('posts b', 'a.postId=b.id', 'LEFT')
+            ->join('user c', 'c.id=a.userId', 'LEFT')
+            ->where([['c.active', 1]])
+            ->orderBy([['a.created_at', 'DESC']])
+            ->get();
+        return $result;
+    }
+
+    public function getUserComments($params)
+    {
+
+        $userId = $params[2];
+        $result = $this->select(['a.*, b.*, c.username, d.profilePic'])
+            ->join('posts b', 'a.postId=b.id')
+            ->join('user c', 'b.userId=c.id')
+            ->join('userinfo d', 'c.id=d.userId')
+            ->where([['a.userId', $userId]])
+            ->orderBy([['b.created_at', 'DESC']])
+            ->get();
+        return $result;
     }
 }
